# optimized brigand::replace and brigand::replace_if