```cpp
using list = brigand::replace<brigand::range<int, 0, SIZE>, int, int>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.11s - 39M | 00.16s - 42M | 00.26s - 49M | 00.48s - 62M
gcc | 00.06s - 21M | 00.11s - 29M | 00.18s - 40M | 00.32s - 61M | 00.60s - 103M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.09s - 38M | 00.12s - 40M | 00.19s - 45M | 00.32s - 54M
gcc | 00.05s - 20M | 00.08s - 25M | 00.13s - 31M | 00.22s - 43M | 00.40s - 67M

---

```cpp
template<class N> struct is_odd { using type = std::integral_constant<bool, (N::value % 2)>; };

using list = brigand::replace<brigand::range<int, 0, SIZE>, is_odd<brigand::_1>, int>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 37M | 00.13s - 40M | 00.20s - 45M | 00.36s - 54M | 00.63s - 73M
gcc | 00.06s - 22M | 00.13s - 33M | 00.22s - 48M | 00.40s - 76M | 00.82s - 133M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.10s - 38M | 00.14s - 41M | 00.22s - 47M | 00.38s - 58M
gcc | 00.05s - 20M | 00.09s - 26M | 00.14s - 34M | 00.25s - 48M | 00.46s - 77M